Task 16. Fibonacci 98

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -8,26 +8,26 @@
 int main(void)
 {
 	int i;
-	unsigned long int n1, n2, sum;
+	double n1, n2, sum;
 
 	n1 = 1;
 	n2 = 2;
 	sum = n1 + n2;
-	printf("%ld, %ld, ", n1, n2);
+	printf("%.0f, %.0f, ", n1, n2);
 
 	i = 3;
 	while (i <= 98)
 	{
 		if (i < 98)
 		{
-			printf("%ld, ", sum);
+			printf("%.0f, ", sum);
 			n1 = n2;
 			n2 = sum;
 			sum = n1 + n2;
 		}
 		else
 		{
-			printf("%ld\n", sum);
+			printf("%.0f\n", sum);
 		}
 		i++;
 	}
